update: move endpoints to a library
