+++ api/shuttle/Cargo.toml
@@ -15,6 +15,8 @@ tokio = "1.26.0"
 
 tracing = "0.1.40"
 
+api-lib = { path = "../lib" }
+
 # env_logger = "0.10.0"
 # log = "0.4.17"
